Move observer and event into replication, move product observers into own module path
