+++ src/lib/entity/mod.rs
@@ -1,0 +1,1 @@
+mod product;
