+++ src/lib/entity/product/mod.rs
@@ -1,0 +1,1 @@
+mod entity;
