+++ src/lib/replication/event/mod.rs
@@ -1,8 +1,7 @@
 use crate::replication::BinaryRow;
-mod observer;
+
 mod update_event;
 
-pub use observer::EventObserver;
 pub use update_event::UpdateEvent;
 
 pub enum Event {
