+++ src/lib/replication/observer/macros.rs
