+++ src/lib/replication/observer/mod.rs
@@ -2,13 +2,12 @@
 mod macros;
 mod chain_observer;
 mod filter_observer;
-mod product_entity;
 
 use crate::error::Error;
-use crate::event::observer::chain_observer::ChainObserver;
-use crate::event::observer::filter_observer::{FilterObserver, FilterObserverPredicate};
-use crate::event::Event;
-use crate::TableSchema;
+use crate::replication::Event;
+use crate::schema::TableSchema;
+use chain_observer::ChainObserver;
+use filter_observer::{FilterObserver, FilterObserverPredicate};
 
 pub trait EventObserver: Sized {
     async fn process_event(&self, event: &Event, table: &impl TableSchema) -> Result<(), Error>;
