+++ src/lib/replication/test_fixture.rs
@@ -8,7 +8,6 @@ use mysql_common::io::ParseBuf;
 use phf::phf_map;
 use phf::Map;
 
-
 static TABLE_DEFINITION_ENTITY: &[u8] = &[
     24, 1, 0, 0, 0, 0, 1, 0, 8, 116, 101, 115, 116, 95, 100, 98, 48, 0, 6, 101, 110, 116, 105, 116,
     121, 0, 4, 3, 15, 252, 246, 5, 252, 3, 2, 12, 4, 14, 1, 1, 128, 2, 3, 252, 255, 0,
@@ -54,11 +53,6 @@ static TABLE_DEFINITION: Map<&'static str, &[u8]> = phf_map! {
     "entity_with_multiple_json" => TABLE_DEFINITION_MULTIPLE_JSON
 };
 
-static PARTIAL_UPDATE_ENTITY_WITH_MULTIPLE_JSON: &[u8] = &[
-    138, 0, 0, 0, 0, 0, 1, 0, 2, 0, 6, 1, 16, 0, 1, 0, 0, 0, 1, 2, 0, 20, 0, 0, 0, 0, 10, 36, 46,
-    102, 101, 97, 116, 117, 114, 101, 100, 7, 12, 5, 102, 97, 108, 115, 101,
-];
-
 static PARTIAL_UPDATE_ENTITY: &[u8] = &[
     89, 0, 0, 0, 0, 0, 1, 0, 2, 0, 4, 1, 2, 0, 2, 0, 0, 0, 0, 0, 17, 0, 65, 119, 101, 115, 111,
     109, 101, 32, 80, 114, 111, 100, 117, 99, 116, 32, 50,
@@ -82,7 +76,6 @@ static DELETE_ENTITY_INT: &[u8] = &[
     111, 0, 0, 0, 0, 0, 1, 0, 2, 0, 5, 255, 224, 2, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0,
     0, 0, 0,
 ];
-static MINIMAL_DELETE_ENTITY_INT: &[u8] = &[138, 0, 0, 0, 0, 0, 1, 0, 2, 0, 5, 1, 254, 2, 0, 0, 0];
 static UPDATE_ROWS_JSON: &[u8] = &[
     111, 0, 0, 0, 0, 0, 1, 0, 2, 0, 5, 255, 255, 224, 1, 0, 0, 0, 2, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0,
     0, 73, 0, 0, 0, 0, 3, 0, 72, 0, 25, 0, 4, 0, 29, 0, 6, 0, 35, 0, 8, 0, 4, 2, 0, 2, 43, 0, 4, 1,
